RBMC: Fill in ReasonsForNoRedundancy D-Bus prop
There is now a 'ReasonsForNoRedundancy' D-Bus property that is a set of
ReasonForNoRedundancy PDI enums.  It needs to be on D-Bus so that the
reasons can be displayed in Redfish.

Switch to returning those enums in the getNoRedundancyReasons()
function, and set the D-Bus property with the result.

The property set itself will now handle writing the values to the
data.json file.  This doesn't need to be done for technical reasons as
redundancy is recalculated on each startup of the active BMC, but it
will make data.json more useful for debug if it's included in dumps or
error logs.

The rbmctool can now just get the reasons from D-Bus when it displays
them instead of data.json.

Since the string versions of the enums are fairly readable, remove the
function that was translating the old enums to a description, and just
use the last segment of the enum string value in data.json and rbmctool
output.

Tested:
Can see the properties on D-Bus:
```
$ busctl get-property xyz.openbmc_project.State.BMC.Redundancy \
         /xyz/openbmc_project/state/bmc0 xyz.openbmc_project.State.BMC.Redundancy ReasonsForNoRedundancy

as 2 "xyz.openbmc_project.State.BMC.Redundancy.ReasonForNoRedundancy.ManuallyDisabled"
     "xyz.openbmc_project.State.BMC.Redundancy.ReasonForNoRedundancy.SiblingNotAtReady"
```

And in rbmctool:
```
Reasons for no BMC redundancy:
    ManuallyDisabled
    SiblingNotAtReady
```

And in data.json:
```
    "NoRedundancyReasons": [
        "ManuallyDisabled",
        "SiblingNotAtReady"
    ],
```

Change-Id: I99172e1033476141a35ff1b7e4504657b320424a
Signed-off-by: Matt Spinler <spinler@us.ibm.com>