RBMC: Simplify isErrorReason function
The isErrorReason() function says if the role reason was one that
required the role to be passive due to an error case.

Simplify that function by keeping the reasons in an array and just
searching that. That list is at the top of the file to make it easier to
remember to add to.

Also add the missing systemInventoryNotAvailable reason.

Tested:

Unit test for this function passes.

Change-Id: I3875adce4c9e94db0bd9d5303a018d0bccf4bead
Signed-off-by: Matt Spinler <spinler@us.ibm.com>